+++ packages/i18n/CHANGELOG.md
@@ -1,5 +1,23 @@
 # gt-i18n
 
+## 0.8.8
+
+### Patch Changes
+
+- [#1278](https://github.com/generaltranslation/gt/pull/1278) [`ce0933a`](https://github.com/generaltranslation/gt/commit/ce0933ab102d34a0c38634f7c2b0d634c9a620a8) Thanks [@bgub](https://github.com/bgub)! - Add a `generaltranslation/core` entrypoint for locale and formatting helpers, and update `gt-i18n` to consume it where possible.
+
+- Updated dependencies [[`ce0933a`](https://github.com/generaltranslation/gt/commit/ce0933ab102d34a0c38634f7c2b0d634c9a620a8)]:
+  - generaltranslation@8.2.9
+  - @generaltranslation/supported-locales@2.0.67
+
+## 0.8.7
+
+### Patch Changes
+
+- Updated dependencies [[`b12d57d`](https://github.com/generaltranslation/gt/commit/b12d57dab1d5cb1f602c5ac24a702b48cda7f11e)]:
+  - generaltranslation@8.2.8
+  - @generaltranslation/supported-locales@2.0.66
+
 ## 0.8.6
 
 ### Patch Changes